add HTTPStatus

+++ quizApp/src/main/java/com/projects/quizApp/service/QuestionService.java
@@ -4,8 +4,11 @@ package com.projects.quizApp.service;
 import com.projects.quizApp.model.Question;
 import com.projects.quizApp.repository.QuestionRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @Service
@@ -14,17 +17,29 @@ public class QuestionService {
 
     @Autowired
     QuestionRepository questionRepository;
-    public List<Question> getAllQuestions() {
-        return questionRepository.findAll();
+
+    /**
+     * Get All Questions from database and return so form Responce Entity with Http Status Ok
+     *
+     */
+    public ResponseEntity<List<Question>> getAllQuestions() {
+        try {
+            return new ResponseEntity<>(questionRepository.findAll(), HttpStatus.OK);
+        }catch(Exception e){
+            e.printStackTrace();
+        }
+        //if something went wrong
+        return new ResponseEntity<>(new ArrayList<>(),HttpStatus.BAD_REQUEST);
     }
     /**
      * save quest
      * @param qst
-     * @return
+     * @return success as message in body request and HTTPSatus Code 201
      */
-    public String saveQuestion(Question qst){
+    public ResponseEntity<String> saveQuestion(Question qst){
         questionRepository.save(qst);
-        return "success";
+        //HTTPSatus Created code : 201
+        return new ResponseEntity<>("success", HttpStatus.CREATED);
     }
     public Question getQuestionById(int id){
         return questionRepository.getById(id);
@@ -41,9 +56,13 @@ public class QuestionService {
         return  questionRepository.getByCategory(category);
     }
 
-    public String deleteQuestion(Integer id) {
+    public ResponseEntity<String> deleteQuestion(Integer id) {
 
-        questionRepository.deleteById(id);
-        return "deleted successful";
+            if(questionRepository.existsById(id)) {
+                questionRepository.deleteById(id);
+                return new ResponseEntity<>("deleted successful", HttpStatus.OK);
+            }
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+
     }
 }
